# internal/api/docs/openapi.yaml
openapi: 3.1.1
info:
  title: FaaS Application API
  description: |
    API for managing and executing serverless functions. This API allows you to create, manage,
    and execute functions written in Lua, with support for versioning, environment variables,
    and execution logging.
  version: 1.0.0
  contact:
    name: FaaS API Support
servers:
  - url: /
    description: Production server
  - url: http://localhost:3000
    description: Local development server
...
security:
  - CookieAuth: []
  - BearerAuth: []
paths:
...
  /api/functions:
    post:
      tags:
        - Functions
      summary: Create a new function
      description: Creates a new function with the provided code and metadata. The first version is automatically created and activated.
      operationId: createFunction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFunctionRequest"
            examples:
              simpleFunction:
                summary: Simple HTTP function
                value:
                  name: hello-world
                  description: A simple hello world function
                  code: |
                    function handler(ctx, event)
                      return {
                        statusCode = 200,
                        body = "Hello, World!",
                        headers = {}
                      }
                    end
      responses:
        "200":
          description: Function created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FunctionWithActiveVersion"
        "400":
          description: Validation error (invalid request body or field constraints violated)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                nameEmpty:
                  summary: Empty name
                  value:
                    error: "name: name cannot be empty"
                nameTooLong:
                  summary: Name too long
                  value:
                    error: "name: name cannot be longer than 100 characters"
                codeEmpty:
                  summary: Empty code
                  value:
                    error: "code: code cannot be empty"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    get:
      tags:
        - Functions
      summary: List all functions
      description: Returns a paginated list of all functions with their active versions
      operationId: listFunctions
      parameters:
        - name: limit
          in: query
          description: Maximum number of items to return (default 20, max 100)
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
            example: 20
        - name: offset
          in: query
          description: Number of items to skip
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
            example: 0
      responses:
        "200":
          description: List of functions retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListFunctionsResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/functions/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: Unique identifier of the function
        schema:
          type: string
          example: "abc123xyz"

    get:
      tags:
        - Functions
      summary: Get a specific function
      description: Returns detailed information about a function including its active version
      operationId: getFunction
      responses:
        "200":
          description: Function retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FunctionWithActiveVersion"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Function not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      tags:
        - Functions
      summary: Update a function
      description: |
        Updates function metadata and/or code. If code is provided, a new version is created.
        All fields are optional.
      operationId: updateFunction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFunctionRequest"
            examples:
              updateMetadata:
                summary: Update only metadata
                value:
                  name: new-function-name
                  description: Updated description
              updateCode:
                summary: Update code (creates new version)
                value:
                  code: |
                    function handler(ctx, event)
                      return {statusCode = 200, body = "Updated!", headers = {}}
                    end
              disableFunction:
                summary: Disable function
                value:
                  disabled: true
              enableFunction:
                summary: Enable function
                value:
                  disabled: false
      responses:
        "200":
          description: Function updated successfully
        "400":
          description: Validation error (invalid request body or field constraints violated)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                noFields:
                  summary: No fields provided
                  value:
                    error: "request: at least one field must be provided for update"
                nameEmpty:
                  summary: Empty name
                  value:
                    error: "name: name cannot be empty"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      tags:
        - Functions
      summary: Delete a function
      description: Permanently deletes a function and all its versions
      operationId: deleteFunction
      responses:
        "204":
          description: Function deleted successfully
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Function not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/functions/{id}/versions:
    parameters:
      - name: id
        in: path
        required: true
        description: Unique identifier of the function
        schema:
          type: string

    get:
      tags:
        - Versions
      summary: List all versions of a function
      description: Returns a paginated list of all versions of a function in descending order (newest first)
      operationId: listVersions
      parameters:
        - name: limit
          in: query
          description: Maximum number of items to return (default 20, max 100)
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
            example: 20
        - name: offset
          in: query
          description: Number of items to skip
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
            example: 0
      responses:
        "200":
          description: Versions retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListVersionsResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/functions/{id}/versions/{version}:
    parameters:
      - name: id
        in: path
        required: true
        description: Unique identifier of the function
        schema:
          type: string
      - name: version
        in: path
        required: true
        description: Version number
        schema:
          type: integer
          minimum: 1
          example: 1

    get:
      tags:
        - Versions
      summary: Get a specific version
      description: Returns detailed information about a specific version of a function
      operationId: getVersion
      responses:
        "200":
          description: Version retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FunctionVersion"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Version not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/functions/{id}/versions/{version}/activate:
    parameters:
      - name: id
        in: path
        required: true
        description: Unique identifier of the function
        schema:
          type: string
      - name: version
        in: path
        required: true
        description: Version number to activate
        schema:
          type: integer
          minimum: 1

    post:
      tags:
        - Versions
      summary: Activate a version
      description: Sets the specified version as the active version for the function
      operationId: activateVersion
      responses:
        "200":
          description: Version activated successfully
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/functions/{id}/executions:
    parameters:
      - name: id
        in: path
        required: true
        description: Unique identifier of the function
        schema:
          type: string

    get:
      tags:
        - Executions
      summary: List executions of a function
      description: Returns a paginated list of execution history for a function
      operationId: listExecutions
      parameters:
        - name: limit
          in: query
          description: Maximum number of items to return (default 20, max 100)
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
            example: 20
        - name: offset
          in: query
          description: Number of items to skip
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
            example: 0
      responses:
        "200":
          description: Executions retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListExecutionsResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
      description: HttpOnly authentication cookie issued after a successful login.
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: APIKey
      description: Provide the API key as a bearer token in the Authorization header.
  schemas:
...
    CreateFunctionRequest:
      type: object
      required:
        - name
        - code
      properties:
        name:
          type: string
          description: Name for the function (must be non-empty after trimming whitespace)
          example: "hello-world"
          minLength: 1
          maxLength: 100
        description:
          type: string
          nullable: true
          description: Optional description
          example: "A simple hello world function"
          maxLength: 500
        code:
          type: string
          description: Lua code for the function (must be non-empty after trimming whitespace)
          example: |
            function handler(ctx, event)
              return {statusCode = 200, body = "Hello", headers = {}}
            end
          minLength: 1
          maxLength: 1048576
...
    FunctionWithActiveVersion:
      allOf:
        - $ref: "#/components/schemas/Function"
        - type: object
          required:
            - active_version
          properties:
            active_version:
              $ref: "#/components/schemas/FunctionVersion"
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
          example: "Function not found"
...